CreateWholeImageFromCBRorCSBS
