+++ acceptance/openstack/csbs/v1/backup_test.go
@@ -3,7 +3,6 @@ package v1
 import (
 	"testing"
 
-	golangsdk "github.com/opentelekomcloud/gophertelekomcloud"
 	"github.com/opentelekomcloud/gophertelekomcloud/acceptance/clients"
 	"github.com/opentelekomcloud/gophertelekomcloud/acceptance/openstack"
 	"github.com/opentelekomcloud/gophertelekomcloud/acceptance/tools"
@@ -45,33 +44,13 @@ func TestBackupLifeCycle(t *testing.T) {
 
 	if query[0].Result {
 		t.Logf("Resource is protectable")
-		backupName := tools.RandomString("backup-", 3)
-
-		t.Logf("Attempting to create CSBS backup")
-		checkpoint, err := backup.Create(client, ecs.ID, backup.CreateOpts{
-			BackupName:   backupName,
-			Description:  "bla-bla",
-			ResourceType: "OS::Nova::Server",
-		})
-		th.AssertNoErr(t, err)
-
-		t.Cleanup(func() {
-			t.Logf("Attempting to delete CSBS backup: %s", checkpoint.Id)
-			err = backup.Delete(client, checkpoint.Id)
-			th.AssertNoErr(t, err)
-
-			err = waitForBackupDeleted(client, 600, checkpoint.Id)
-			th.AssertNoErr(t, err)
-			t.Logf("Deleted CSBS backup: %s", checkpoint.Id)
-		})
+		checkpoint := CreateCSBS(t, client, ecs.ID)
 
 		csbsBackupList, err := backup.List(client, backup.ListOpts{
 			CheckpointId: checkpoint.Id,
 		})
 		th.AssertNoErr(t, err)
 
-		err = waitForBackupCreated(client, 600, csbsBackupList[0].Id)
-		th.AssertNoErr(t, err)
 		t.Logf("Created CSBS backup: %s", checkpoint.Id)
 
 		capabilities, err := resource.GetResRestorationCapabilities(client, []resource.GetRestorationOpts{
@@ -94,36 +73,4 @@ func TestBackupLifeCycle(t *testing.T) {
 	} else {
 		t.Logf("Resource isn't protectable")
 	}
-}
-
-func waitForBackupCreated(client *golangsdk.ServiceClient, secs int, backupID string) error {
-	return golangsdk.WaitFor(secs, func() (bool, error) {
-		csbsBackup, err := backup.Get(client, backupID)
-		if err != nil {
-			return false, err
-		}
-
-		if csbsBackup.Id == "error" {
-			return false, err
-		}
-
-		if csbsBackup.Status == "available" {
-			return true, nil
-		}
-
-		return false, nil
-	})
-}
-
-func waitForBackupDeleted(client *golangsdk.ServiceClient, secs int, backupID string) error {
-	return golangsdk.WaitFor(secs, func() (bool, error) {
-		_, err := backup.Get(client, backupID)
-		if err != nil {
-			if _, ok := err.(golangsdk.ErrDefault404); ok {
-				return true, nil
-			}
-			return false, err
-		}
-		return false, nil
-	})
 }
